requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
accelerate==1.5.2
    # via -r requirements.in
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
diffusers==0.32.2
    # via -r requirements.in
filelock==3.18.0
    # via
    #   -r requirements.in
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.3.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.29.3
    # via
    #   accelerate
    #   diffusers
    #   optimum-quanto
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
importlib-metadata==8.6.1
    # via diffusers
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
ninja==1.11.1.3
    # via optimum-quanto
numpy==2.0.2
    # via
    #   accelerate
    #   diffusers
    #   optimum-quanto
    #   torchvision
    #   transformers
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
optimum-quanto==0.2.7
    # via -r requirements.in
packaging==24.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
pillow==11.1.0
    # via
    #   -r requirements.in
    #   diffusers
    #   torchvision
protobuf==6.30.1
    # via -r requirements.in
psutil==7.0.0
    # via accelerate
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.3
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via
    #   -r requirements.in
    #   accelerate
    #   diffusers
    #   optimum-quanto
    #   transformers
sentencepiece==0.2.0
    # via -r requirements.in
sympy==1.13.1
    # via torch
tokenizers==0.21.1
    # via transformers
torch==2.6.0
    # via
    #   -r requirements.in
    #   accelerate
    #   optimum-quanto
    #   torchvision
torchvision==0.21.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.49.0
    # via -r requirements.in
triton==3.2.0
    # via torch
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   torch
urllib3==2.3.0
    # via requests
zipp==3.21.0
    # via importlib-metadata
